Separate progress monitoring from VoltDBEngine 2
